+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.15.0 (27-01-2024)
+
+-   `WrongOverrideLinter` now checks if member is deprecated
+-   Fix: `UnusedClassLinter` doesn't show error if any member is used outside the class
+
 ## 1.14.5 (25-01-2024)
 
 -   [UI5 Parser](https://github.com/iljapostnovs/ui5plugin-parser) updated to v1.7.7